Remove the experimental tasty-reader for scala 2
The akka module is not published for scala 3. If the root project
defines all `crossScalaVersions` then the akka_2.13 subproject is
compiled against the core_3 subproject; but since the core also exists
for 2.13, this is not necessary. Setting the `crossScalaVersions` key
to `Nil` for the root project allows to remove the `-Ytasty-reader`
setting; sbt compiles the akka suproject against core_2.13.